LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_mark_with_name_of_navigation_timing_optional_attribute-expected.txt

CONSOLE MESSAGE: line 58: ReferenceError: Can't find variable: wp_test
CONSOLE MESSAGE: line 30: ReferenceError: Can't find variable: test_true

Harness Error (FAIL), message = ReferenceError: Can't find variable: wp_test

Description

This test validates exception scenarios of invoking performance.mark() with param of "secureConnectionStart".
